+++ BP/scripts/Core/main.js
@@ -1,6 +1,0 @@
-import '../DoriosLib/index.js'
-
-import './stats_manager.js'
-import './update_stats.js'
-import './active_abilities.js'
-import './loot_injector.js'