+++ BP/scripts/DoriosRPGCore/main.js
@@ -1,0 +1,7 @@
+import '../DoriosLib/index.js'
+
+import './scriptEvents.js'
+import './statsManager.js'
+import './updateStats.js'
+import './activeAbilities.js'
+import './lootInjector.js'